+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 1. Added `SetDoorPasscodes` API function to set door supervisor passcodes.
 
+### Updates
+1. Reworked `nil` Event pointer in `GetStatus` as a zero value.
+
 
 ## [0.8.6](https://github.com/uhppoted/uhppoted-lib/releases/tag/v0.8.6) - 2023-08-30